Removed MdProcessor file, and updated README, Program.cs and HtmlProcessor.cs to match owner comments
